requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
redis==5.0.1
python-dotenv==1.0.0
websockets==11.0.3
python-multipart==0.0.6
pyyaml==6.0.2